Changed: Made Multiple Changes
Changed the folder structure and general directory layout for easier installation purposes.

Updated the ReadMe file to be more concise and for easier readability.

Added a minimum API version check, to ensure that the server is running at least version 96 of CounterStrikeSharp when using the plugin.

Implemented a Windows an Linux server operating system check, thanks to KillStreak for this.

Made changes to how the altering of the player's collision data is being handled.